@@ -4,7 +4,7 @@ description: An example project for persistent_bottom_nav_bar_v2
 version: 1.0.0
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
 
 dependencies:
   flutter:
